LayoutTests/http/tests/loading/preload-ignore-invalid-base.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<script>
window.fail = false;
</script>
<base href="gopher:???:"></base>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=100"></script>
<script>
shouldBeFalse('internals.isPreloaded("resources/fail.js")')
</script>
<!-- The resource should not be read, as the baseUrl is now set to empty from reading invalid url -->
<script src="resources/fail.js"></script>
<script>
shouldBeFalse('window.fail')
</script>
