LayoutTests/http/tests/loading/preload-ignore-invalid-base-expected.txt

PASS internals.isPreloaded("resources/fail.js") is false
PASS window.fail is false
PASS successfullyParsed is true

TEST COMPLETE
